backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0
ollama>=0.1.8
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0 # Explicitly add requests
httpx>=0.27.0 # For ollama client
websockets>=12.0 # Add websockets

# Agents & Workflow

# File Handling
pandas
openpyxl
python-multipart # Required by FastAPI for file uploads

# LLM Interaction
